SCRD-8072 Cleanup ardana_notify restart handling (SCRD-8072)
The existing code doesn't correctly leverage the ardana_notify_...
variables to determine if the relevant manila service components
need to be started.

As a result it may unnecessarily restart the manila-api service
component, and may never restart the manila-scheduler service
component. This is addressed by correctly making use of the
persisted ardana_notify_... registered facts to determine if
system state has changed and needs to be updated. This also
requires that we "sanely" clear the persisted state for these
facts in the ansible fact cache if we have been successful in
restarting the services.

One caveat with the new implementation is that we potentially
could unnecessarily restart the service components if the start
task is included in an ansible-playbook play more than once, e.g
indirectly due to including some other play. If this is indeed
a concern it can be rectified by splitting the restarting of
each of the service components into explicit independent actions,
rather than using the with_items iteration that we currently
make use of.

Additionally this patch also removes the unnecessary additional
templating the 100-manila.conf file in the MNL-API configure.yml
task, after having done so in the included _MNL-CMN configure.yml
task.

We split the manila-start.yml up into two sub-plays that we can
include as needed elsewhere to ensure that we don't unnecessarily
clear the ardana_notify_manila_restart_required fact too early
in the case of a manila-reconfigure.yml run.

Change-Id: Ia0fa9f8c7fb5c652c171998b1894e55f25a39bc2