SCRD-7414 Setup audit filter for manila (SCRD-7414)
While we are correctly enabling the auditing mechanism for Manila at a
high level, no audit entries are being captured for Manila because we
don't actually configure the audit plugin in the api_paste.ini for the
manila-api service, nor are we generating the required audit_map.conf
file.

This patch addresses both of these issues, so that we do capture audit
log entries for manila-api.

However we need to flesh out the api_audit_map.conf.j2 more fully; the
upstream PyCADF project doesn't have a reference example for Manila.

Change-Id: Ic9d9c39711faf38f1c666d2378eeb82778a283a7