Manila parallelised upgrade workflow enhancements (SOC-10609)
The Cloud8 Manila deployment is RPM, rather than VENV, based, and the
openstack-manila-api service is configured to run the manila-manage
db sync command as a pre-start action. This means that we need to
carefully orchestrate when the openstack-manila-api is started to
avoid any risk of database corruption due to multiple nodes starting
the service at the same time.

However the Cloud8 Manila deployment may be setting the installed
services to be enabled in systemd, meaning that they automatically
restart on failure or RPM package upgrade. This poses a problem for
the parallelised upgrade where the nodes are zypper dist-upgraded
potentially at the same time, meaning that two instances of the
openstack-manila-api could be triggered to restart at approximately
the same time.

To address this we need to disable the automatic restart of the
services in systemd which is handled in the new manila-pre-upgrade.yml
playbook, intended to be run by the ardana-upgrade.yml playbook before
the package update step is triggered.

Note that this solution cannot avoid the automatic restart of any
Manila services deployed on a controller that is also the deployer,
i.e. integrated deployer scenario, because the zypper dist-upgrade
will have happened on the deployer node before we ever get to run
the ardana-upgrade.yml. However only a single node will be affected
so we shouldn't have to worry about multiple nodes trying to run
the manila-manage db sync command in parallel... But if the db sync
makes significant changes to the Manila DB format, we still have to
deal with the risk of database corruption if the DB format gets
upgraded by the new Cloud9 based instance of openstack-manila-api
while there are still Cloud8 based versions running; this will need
to be addressed by changes in the Cloud8 stream to disable the
automatic restart of the services when deployed/updated.

Renamed manila_component fact to an appropriate role specific name
so as to avoid issues in high-level playbook run contexts due to
the MNL-API and MNL-SHR roles getting loaded multiple times, which
lead to incorrect values being used by subsequent MNL-API targetted
plays if MNL-SHR had been loaded after MNL-API.

Also added support for detecting if any manila shares have actually
been configured to avoid reporting failure for non-existing manila
share service instances.

Change-Id: Ie7afce842d29d9afdd7ae2143392ef72ef8b2933