Skip openstack-manila-share status check during upgrade (SOC-11341)
After a Cloud8 system has been successfully updated and rebooted by the
ardana-update.yml workflow, the openstack-manila-share services may not
be running; this differs from the state of a fresh deployment of Cloud8
that we have been testintg the ardana-upgrade.yml workflow against.

To workaround failing the ardana-upgrade.yml workflow for this reason
we can apply the same workaround that was already used for the other
Manila services during the pre-upgrade phase of the upgrade, namely to
skip the unreliable status check.

Change-Id: I83fb80cd51c02d65c7807d44b0651cc5a4fafcb5